refactor/rename HasTag to XMLTagShield
